requirements.txt
aiogram>=3.0,<4.0
playwright>=1.40
aiosqlite>=0.19
cryptography>=41.0
python-dotenv>=1.0
beautifulsoup4>=4.12
